+++ src/functions/public/Add-PsModule.ps1
@@ -1,0 +1,65 @@
+function Add-PsModule {
+    <#
+    .Synopsis
+        Import a Module in a [System.Management.Automation.Runspaces.InitialSessionState] instance
+    .Description
+        Used by Add-PsCommand
+        Uses ImportPSModule() or ImportPSModulesFromPath() depending on the module
+    .EXAMPLE
+        $InitialSessionState = [system.management.automation.runspaces.initialsessionstate]::CreateDefault()
+        Add-PsModule -InitialSessionState $InitialSessionState -ModuleInfo $ModuleInfo
+    #>
+
+    param(
+
+        # Powershell interface to add the Command to
+        [Parameter(Mandatory)]
+        [System.Management.Automation.Runspaces.InitialSessionState]$InitialSessionState,
+
+        <#
+        ModuleInfo object for the module to add to the Powershell interface
+        #>
+        [Parameter(
+            Mandatory,
+            Position = 0
+        )]
+        [System.Management.Automation.PSModuleInfo[]]$ModuleInfo
+
+    )
+
+    begin {
+
+        $TodaysHostname = HOSTNAME.EXE
+
+    }
+
+    process {
+
+        ForEach ($ThisModule in $ModuleInfo) {
+
+            switch ($ThisModule.ModuleType) {
+                'Binary' {
+                   <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tAdd-PsModule`t`$InitialSessionState.ImportPSModule('$($ThisModule.Name)')"
+                    $InitialSessionState.ImportPSModule($ThisModule.Name)
+                }
+                'Script' {
+                    $ModulePath = Split-Path -Path $ThisModule.Path -Parent
+                   <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tAdd-PsModule`t`$InitialSessionState.ImportPSModulesFromPath('$ModulePath')"
+                    $InitialSessionState.ImportPSModulesFromPath($ModulePath)
+                }
+                'Manifest' {
+                    $ModulePath = Split-Path -Path $ThisModule.Path -Parent
+                   <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tAdd-PsModule`t`$InitialSessionState.ImportPSModulesFromPath('$ModulePath')"
+                    $InitialSessionState.ImportPSModulesFromPath($ModulePath)
+                }
+                default {
+                    # Scriptblocks or Functions not from modules will have no module to import so ModuleInfo will be null
+                }
+
+            }
+
+        }
+
+    }
+
+}