+++ src/functions/public/Split-Thread.ps1
@@ -3,37 +3,29 @@ function Split-Thread {
     <#
     .Synopsis
         Split a command for a collection of input objects into multiple threads for asynchronous processing
-
     .Description
         The specified command will be run for each input object in a separate powershell instance with its own runspace
         These runspaces are part of the same runspace pool inside the same powershell.exe process
-
     .EXAMPLE
         The following demonstrates sending a Cmdlet name to the -Command parameter
             $InputObject | Split-Thread -Command 'Write-Output'
-
     .EXAMPLE
         The following demonstrates sending a scriptblock to the -Command parameter
             $InputObject | Split-Thread -Command [scriptblock]::create("Write-Output `$args[0]")
-
     .EXAMPLE
         The following demonstrates sending a script file path to the -Command parameter
             $InputObject | Split-Thread -Command "C:\Test-Command.ps1"
-
     .EXAMPLE
         The following demonstrates sending a function to the -Command parameter
             $InputObject | Split-Thread -Command 'Test-Function'
-
     .EXAMPLE
         The following demonstrates the -AddParam parameter
 
         $InputObject | Split-Thread -Command "Get-Service" -InputParameter ComputerName -AddParam @{"Name" = "BITS"}
-
     .EXAMPLE
         The following demonstrates the -AddSwitch parameter
 
         $InputObject | Split-Thread -Command "Get-Service" -AddSwitch @('RequiredServices','DependentServices')
-
 	.EXAMPLE
 		The following demonstrates the use of a threadsafe hashtable to store results
 		The hastable can be accessed and updated from inside each runspace
@@ -92,48 +84,39 @@ function Split-Thread {
 
     begin {
 
-        $InitialSessionState = [system.management.automation.runspaces.initialsessionstate]::CreateDefault()
+        $TodaysHostname = HOSTNAME.EXE
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Entered begin block for '$Command'"
+
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t`$InitialSessionState = [System.Management.Automation.Runspaces.InitialSessionState]::CreateDefault() # for '$Command'"
+        $InitialSessionState = [System.Management.Automation.Runspaces.InitialSessionState]::CreateDefault()
 
         # Import the source module containing the specified Command in each thread
-        $CommandInfo = Get-PsCommandInfo -Command $Command
-        switch ($CommandInfo.ModuleInfo.ModuleType) {
-            'Binary' {
-                Write-Debug "`$InitialSessionState.ImportPSModule('$($CommandInfo.ModuleInfo.Name)')"
-                $InitialSessionState.ImportPSModule($CommandInfo.ModuleInfo.Name)
-            }
-            'Script' {
-                $ModulePath = $CommandInfo.ModuleInfo.Path | Split-Path -Parent
-                Write-Debug "`$InitialSessionState.ImportPSModulesFromPath('$ModulePath')"
-                $InitialSessionState.ImportPSModulesFromPath($ModulePath)
-            }
-            'Manifest' {
-                $ModulePath = $CommandInfo.ModuleInfo.Path | Split-Path -Parent
-                Write-Debug "`$InitialSessionState.ImportPSModulesFromPath('$ModulePath')"
-                $InitialSessionState.ImportPSModulesFromPath($ModulePath)
-            }
-            default {
-                # Scriptblocks have no module to import so ModuleInfo will be null
-            }
-        }
 
-        # Import any additional specified modules in each thread
+        $OriginalCommandInfo = Get-PsCommandInfo -Command $Command
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Found 1 original PsCommandInfo for '$Command'"
+        $CommandInfo = Expand-PsCommandInfo -PsCommandInfo $OriginalCommandInfo
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Found $(($CommandInfo | Measure-Object).Count) nested PsCommandInfos for '$Command' ($($CommandInfo.CommandInfo.Name -join ','))"
+
+        # Prepare our collection of PowerShell modules to import in each thread
+        # This will include any modules specified by name with the -AddModule parameter
+        # This will also include any modules identified by tokenizing the -Command parameter or its definition, and recursing through all nested command tokens
+        $ModulesToAdd = [System.Collections.Generic.List[System.Management.Automation.PSModuleInfo]]::new()
         ForEach ($Module in $AddModule) {
-
             $ModuleObj = Get-Module $Module -ErrorAction SilentlyContinue
-            switch ($ModuleObj.ModuleType) {
-                'Binary' {
-                    Write-Debug "`$InitialSessionState.ImportPSModule('$Module')"
-                    $InitialSessionState.ImportPSModule($Module)
-                }
-                default {
-                    # This is for Script or Manifest modules
-                    Write-Debug "`$InitialSessionState.ImportPSModulesFromPath('$($ModuleObj.Path | Split-Path -Parent)')"
-                    $InitialSessionState.ImportPSModulesFromPath(($ModuleObj.Path | Split-Path -Parent))
-
-                }
-            }
-
+            $null = $ModulesToAdd.Add($ModuleObj)
         }
+
+        $ModulesToAdd = $CommandInfo.ModuleInfo |
+        Sort-Object -Property Name -Unique
+
+        $CommandInfo = $CommandInfo |
+        Where-Object -FilterScript {
+            $ModulesToAdd.Name -notcontains $_.ModuleInfo.Name -and
+            $_.CommandType -ne 'Cmdlet'
+        }
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Found $(($CommandInfo | Measure-Object).Count) remaining PsCommandInfos to define for '$Command' (not in modules: $($CommandInfo.CommandInfo.Name -join ','))"
+
+        $null = Add-PsModule -InitialSessionState $InitialSessionState -ModuleInfo $ModulesToAdd
 
         # Set the preference variables for PowerShell output streams in each thread to match the current preferences
         $OutputStream = @('Debug', 'Verbose', 'Information', 'Warning', 'Error')
@@ -148,8 +131,10 @@ function Split-Thread {
             $InitialSessionState.Variables.Add($VariableEntry)
         }
 
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t`$RunspacePool = [runspacefactory]::CreateRunspacePool(1, $Threads, `$InitialSessionState, `$Host) # for '$Command'"
         $RunspacePool = [runspacefactory]::CreateRunspacePool(1, $Threads, $InitialSessionState, $Host)
-        $VerbosePreference = 'SilentlyContinue'
+        #####don'trememberwhythisishere#####$VerbosePreference = 'SilentlyContinue'
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t`$RunspacePool.Open() # for '$Command'"
         $RunspacePool.Open()
 
         $Global:TimedOut = $false
@@ -159,6 +144,12 @@ function Split-Thread {
     }
 
     process {
+        if ($ObjectStringProperty) {
+            $ObjectString = $InputObject.$ObjectStringProperty
+        } else {
+            $ObjectString = $InputObject.ToString()
+        }
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Entered process block for '$Command' on '$ObjectString'"
 
         # Add all the input objects from the pipeline to a single collection; allows progress bars later
         ForEach ($ThisObject in $InputObject) {
@@ -167,6 +158,8 @@ function Split-Thread {
 
     }
     end {
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Entered end block for '$Command'"
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Sending $(($CommandInfo | Measure-Object).Count) PsCommandInfos to Open-Thread for '$Command'"
         $ThreadParameters = @{
             Command              = $Command
             InputParameter       = $InputParameter
@@ -178,18 +171,19 @@ function Split-Thread {
             RunspacePool         = $RunspacePool
         }
         $AllThreads = Open-Thread @ThreadParameters
+        <#NormallyCommentThisForPerformanceOptimization#>#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t# Received $(($AllThreads | Measure-Object).Count) threads from Open-Thread for $Command"
         Wait-Thread -Thread $AllThreads -Threads $Threads -SleepTimer $SleepTimer -Timeout $Timeout -Dispose
         $VerbosePreference = 'Continue'
 
         if ($Global:TimedOut -eq $false) {
 
-            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$(hostname)`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Close()"
+            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Close()"
             $null = $RunspacePool.Close()
-            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$(hostname)`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Close() completed"
+            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Close() completed"
 
-            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$(hostname)`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Dispose()"
+            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Dispose()"
             $null = $RunspacePool.Dispose()
-            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$(hostname)`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Dispose() completed"
+            #CommentedForPerformanceOptimization#Write-Debug "  $(Get-Date -Format s)`t$TodaysHostname`tSplit-Thread`t[System.Management.Automation.Runspaces.RunspacePool]::Dispose() completed"
 
         }